Переход на конкретного пользователя

+++ src/main/java/ru/sertok/spring/jdbc/dao/api/CrudDao.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 
 public interface CrudDao<T> {
     void save(T model);
-    Optional<T> find(Integer id);
+    Optional<T> find(Long id);
     void update(T model);
-    void delete(Integer id);
+    void delete(Long id);
     List<T> findAll();
 }
